+++ src/booking/entities/booking.ts
@@ -1,4 +1,4 @@
-import { Room } from 'src/room/entities/room.en';
+import { Room } from 'src/room/entities/room';
 import { User } from 'src/user/entities/user';
 import { Column, Entity, ManyToOne, PrimaryGeneratedColumn } from 'typeorm';
