+++ src/hotel/dto/get-hotels.ts
