+++ src/room/entities/room.ts
